// src/interfaces/student.interface.ts
export interface EmployeeInt {
  first_name: string;
  last_name: string;
  email: string;
  password: string;
  gender: string;
  dob: string;
  nationality: string;
  highest_qualifications: string;
  phone: string;
  department: string;
  snnit_no: string;
  tin: string;
}

export interface User {
  name: string;
  email: string;
  password: string;
}